Changing pipelines.py and renaming main.py to __main__.py
