+++ kubeflow/pipelines.py
@@ -102,6 +102,7 @@ def train_op(
             "-m",
             "torch.distributed.launch",
             f"--nproc_per_node={num_gpu}",
+            "-m",
             "datasetinsights",
             "train",
         ]
